% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compareMV}
\alias{compareMV}
\title{Compare columns and vector}
\usage{
compareMV(x, y, s = "-")
}
\arguments{
\item{x}{nxm character matrix}

\item{y}{character vector of length n}

\item{s}{string denoting the gaps in matrix and vector that
will not be used for comparison}
}
\value{
Numeric matrix with a relative (regarding to the gap string)
best match for the column
in column one and a relative best match for the vector in column 2.
}
\description{
Matches columns of a matrix against a vector.
}
