% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Matrix multiplication}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A}{numeric nxm matrix}

\item{B}{numeric mxl matrix}
}
\value{
numeric nxl matrix
}
\description{
Matrix multiplication
}
