% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extendData}
\alias{extendData}
\title{Extend data by degree 2}
\usage{
extendData(x, exp = 0.5, base = 2, fac = 1)
}
\arguments{
\item{x}{numeric nxm matrix with variables as columns}

\item{exp}{exponent (see polynomial kernel)}

\item{base}{basis (see polynomial kernal)}

\item{fac}{factor (see polynomial kernel)}
}
\value{
extended matrix with ((n-1)*n)/2 + 2n columns
}
\description{
All interaction terms (including quadratics) will be computed.
}
