% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hcipp}
\alias{hcipp}
\title{hcipp}
\usage{
hcipp(x, y, c)
}
\arguments{
\item{x}{vector of the continuous groud truth}

\item{y}{vector of the predicted values}

\item{c}{vector indicating censored (0) or uncensored (1) data}
}
\value{
h vector of length three with number of
concordances, discordances and the index
}
\description{
Harrelson's concordance index
}
