% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transClose_Del}
\alias{transClose_Del}
\title{Transitive closure after deletion}
\usage{
transClose_Del(x, u, v)
}
\arguments{
\item{x}{numeric nxn matrix (should be binary) that was transitively closed before an edge deletion}

\item{u}{row index of deleted edge}

\item{v}{column index of deleted edge}
}
\value{
transitive closure of x
}
\description{
Transitive closure after deletion
}
