% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transClose_Ins}
\alias{transClose_Ins}
\title{Transitive closure after addition}
\usage{
transClose_Ins(x, u, v)
}
\arguments{
\item{x}{numeric nxn matrix (should be binary) that was transitively closed before an edge addition}

\item{u}{row index of added edge}

\item{v}{column index of added edge}
}
\value{
transitive closure of x
}
\description{
Transitive closure after addition
}
