% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transClose_W}
\alias{transClose_W}
\title{Transitive closure}
\usage{
transClose_W(x)
}
\arguments{
\item{x}{numeric nxn matrix (should be binary)}
}
\value{
transitive closure of x
}
\description{
Transitive closure
}
