% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichExtreme}
\alias{whichExtreme}
\title{whichExtreme}
\usage{
whichExtreme(m, type = 1L, extreme = 1L)
}
\arguments{
\item{m}{rxc matrix}

\item{type}{1 for row and 2 for column indices}

\item{extreme}{1 for minimum and 2 for maximum}
}
\value{
vector of indices
}
\description{
Get the index of the minimum/maximum for each row/column
of a matrix
}
