% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extendData}
\alias{extendData}
\title{Extend data by degree 2}
\usage{
extendData(x, exp = 0.5, base = 2, coef0 = 1)
}
\arguments{
\item{x}{numeric nxm matrix with variables as columns}

\item{exp}{exponent (see polynomial kernel)}

\item{base}{base (see polynomial kernel)}

\item{coef0}{factor (see polynomial kernel)}
}
\value{
extended matrix with ((n-1)*n)/2 + 2n + 1 columns
}
\description{
All interaction terms (including quadratic) will be computed.
}
