% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxCol_row}
\alias{maxCol_row}
\title{Maximum per row}
\usage{
maxCol_row(x)
}
\arguments{
\item{x}{numeric nxm matrix}
}
\value{
numeric vector of length n with the maximum for each row
}
\description{
Maximum per row
}
